Struct cumulus_relay_chain_interface::CommittedCandidateReceipt

pub struct CommittedCandidateReceipt<H = H256> {
    pub descriptor: CandidateDescriptor<H>,
    pub commitments: CandidateCommitments,
}

A candidate-receipt with commitments directly included.

Fields
descriptor: CandidateDescriptor<H>

The descriptor of the candidate.

commitments: CandidateCommitments

The commitments of the candidate receipt.

Implementations

impl<H> CommittedCandidateReceipt<H>

pub fn descriptor(&self) -> &CandidateDescriptor<H>

Get a reference to the candidate descriptor.

impl<H> CommittedCandidateReceipt<H>
where H: Clone,

pub fn to_plain(&self) -> CandidateReceipt<H>

Transforms this into a plain CandidateReceipt.

pub fn hash(&self) -> CandidateHash
where H: Encode,

Computes the hash of the committed candidate receipt.

This computes the canonical hash, not the hash of the directly encoded data. Thus this is a shortcut for candidate.to_plain().hash().

pub fn corresponds_to(&self, receipt: &CandidateReceipt<H>) -> bool
where H: PartialEq,

Does this committed candidate receipt corresponds to the given CandidateReceipt?