Crate cumulus_relay_chain_interface

Structs

CommittedCandidateReceipt

A candidate-receipt with commitments directly included.

InboundDownwardMessage

A wrapped version of DownwardMessage. The difference is that it has attached the block number when the message was sent.

InboundHrmpMessage

An HRMP message seen from the perspective of a recipient.

OverseerHandle

A handle used to communicate with the Overseer.

ParaId

Unique identifier of a parachain.

PersistedValidationData

The validation data provides information about how to create the inputs for validation of a candidate. This information is derived from the chain state and will vary from para to para, although some fields may be the same for every para.

Enums

OccupiedCoreAssumption

An assumption being made about the state of an occupied core.

RelayChainError

Traits

RelayChainInterface

Trait that provides all necessary methods for interaction between collator and relay chain.

Type Aliases

PHash

A hash of some data used by the relay chain.

PHeader

Header type.

RelayChainResult

SessionIndex

Simple index type with which we can count sessions.

StorageValue

Storage value.

ValidatorId

Identity that parachain validators use when signing validation messages.