Trait cumulus_relay_chain_interface::RelayChainInterface

pub trait RelayChainInterface: Send + Sync {

    fn get_storage_by_key(
        &self, 
        block_id: &BlockId, 
        key: &[u8]
    ) -> Result<Option<StorageValue>, Error>;

    fn validators(
        &self, 
        block_id: &BlockId
    ) -> Result<Vec<ValidatorId>, ApiError>;

    fn block_status(&self, block_id: BlockId) -> Result<BlockStatus, Error>;

    fn best_block_hash(&self) -> PHash;

    fn retrieve_dmq_contents(
        &self, 
        para_id: ParaId, 
        relay_parent: PHash
    ) -> Option<Vec<InboundDownwardMessage>>;

    fn retrieve_all_inbound_hrmp_channel_contents(
        &self, 
        para_id: ParaId, 
        relay_parent: PHash
    ) -> Option<BTreeMap<ParaId, Vec<InboundHrmpMessage>>>;

    fn persisted_validation_data(
        &self, 
        block_id: &BlockId, 
        para_id: ParaId, 
        _: OccupiedCoreAssumption
    ) -> Result<Option<PersistedValidationData>, ApiError>;

    fn candidate_pending_availability(
        &self, 
        block_id: &BlockId, 
        para_id: ParaId
    ) -> Result<Option<CommittedCandidateReceipt>, ApiError>;

    fn session_index_for_child(
        &self, 
        block_id: &BlockId
    ) -> Result<SessionIndex, ApiError>;

    fn import_notification_stream(&self) -> ImportNotifications<PBlock>;

    fn wait_for_block<'life0, 'async_trait>(
        &'life0 self, 
        hash: PHash
    ) -> Pin<Box<dyn Future<Output = Result<(), WaitError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn finality_notification_stream(&self) -> FinalityNotifications<PBlock>;

    fn storage_changes_notification_stream(
        &self, 
        filter_keys: Option<&[StorageKey]>, 
        child_filter_keys: Option<&[(StorageKey, Option<Vec<StorageKey>>)]>
    ) -> Result<StorageEventStream<PHash>>;

    fn is_major_syncing(&self) -> bool;

    fn overseer_handle(&self) -> Option<OverseerHandle>;

    fn prove_read(
        &self, 
        block_id: &BlockId, 
        relevant_keys: &Vec<Vec<u8>>
    ) -> Result<Option<StorageProof>, Box<dyn Error>>;

}

Trait that provides all necessary methods for interaction between collator and relay chain.

Required methods

fn get_storage_by_key(
&self,
block_id: &BlockId,
key: &[u8]
) -> Result<Option<StorageValue>, Error>

Fetch a storage item by key.

fn validators(&self, block_id: &BlockId) -> Result<Vec<ValidatorId>, ApiError>

Fetch a vector of current validators.

fn block_status(&self, block_id: BlockId) -> Result<BlockStatus, Error>

Get the status of a given block.

fn best_block_hash(&self) -> PHash

Get the hash of the current best block.

fn retrieve_dmq_contents(
&self,
para_id: ParaId,
relay_parent: PHash
) -> Option<Vec<InboundDownwardMessage>>

Returns the whole contents of the downward message queue for the parachain we are collating for.

Returns None in case of an error.

fn retrieve_all_inbound_hrmp_channel_contents(
&self,
para_id: ParaId,
relay_parent: PHash
) -> Option<BTreeMap<ParaId, Vec<InboundHrmpMessage>>>

Returns channels contents for each inbound HRMP channel addressed to the parachain we are collating for.

Empty channels are also included.

fn persisted_validation_data(
&self,
block_id: &BlockId,
para_id: ParaId,
_: OccupiedCoreAssumption
) -> Result<Option<PersistedValidationData>, ApiError>

Yields the persisted validation data for the given ParaId along with an assumption that should be used if the para currently occupies a core.

Returns None if either the para is not registered or the assumption is Freed and the para already occupies a core.

fn candidate_pending_availability(
&self,
block_id: &BlockId,
para_id: ParaId
) -> Result<Option<CommittedCandidateReceipt>, ApiError>

Get the receipt of a candidate pending availability. This returns Some for any paras assigned to occupied cores in availability_cores and None otherwise.

fn session_index_for_child(
&self,
block_id: &BlockId
) -> Result<SessionIndex, ApiError>

Returns the session index expected at a child of the block.

fn import_notification_stream(&self) -> ImportNotifications<PBlock>

Get a stream of import block notifications.

fn wait_for_block<'life0, 'async_trait>(
&'life0 self,
hash: PHash
) -> Pin<Box<dyn Future<Output = Result<(), WaitError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Wait for a block with a given hash in the relay chain.

This method returns immediately on error or if the block is already reported to be in chain. Otherwise, it waits for the block to arrive.

fn finality_notification_stream(&self) -> FinalityNotifications<PBlock>

Get a stream of finality notifications.

fn storage_changes_notification_stream(
&self,
filter_keys: Option<&[StorageKey]>,
child_filter_keys: Option<&[(StorageKey, Option<Vec<StorageKey>>)]>
) -> Result<StorageEventStream<PHash>>

Get a stream of storage change notifications.

fn is_major_syncing(&self) -> bool

Whether the synchronization service is undergoing major sync. Returns true if so.

fn overseer_handle(&self) -> Option<OverseerHandle>

Get a handle to the overseer.

fn prove_read(
&self,
block_id: &BlockId,
relevant_keys: &Vec<Vec<u8>>
) -> Result<Option<StorageProof>, Box<dyn Error>>

Generate a storage read proof.